Move context into a module
